# docs/swagger.yaml
definitions:
...
  internal_controllers.ChangePasswordRequest:
    properties:
      new_password:
        type: string
    type: object
...
info:
  contact: {}
  description: Сервис для загрузки документов, подсчёта TF‑IDF и управления коллекциями.
  title: LestaStartTest API
  version: 2.2.1
paths:
...
  /user/{user_id}:
    delete:
      description: Удаляет пользователя, его документы и коллекции.
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"User deleted"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление пользователя
      tags:
      - Пользователь
    patch:
      consumes:
      - application/json
      description: Обновляет пароль текущего пользователя.
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: integer
      - description: Новый пароль
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"Password updated"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Password update failed or database error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменение пароля пользователя
      tags:
      - Пользователь
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"